backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.20
pydantic-settings==2.7.0
httpx==0.28.1
numpy==2.2.1
pandas==2.2.3
scikit-learn==1.6.0
xgboost==2.1.3
joblib==1.4.2
apscheduler==3.10.4
razorpay==1.4.2
feedparser==6.0.11
celery==5.4.0
redis==5.0.3
